Fix(smart-crop): Ensure that matches are all highlighted
In the case of only 1 query word was matching in an attribute,
and this same word was matching several times in the crop window of same attribute,
then only the first match was highlighted.
Now we ensure that the computed "best matches interval" contains all the words that match in the same crop window.

related to https://github.com/meilisearch/meilisearch/issues/2627